refactor(ai-gen): split ai_generation.py into ai_generation/ subpackage (Related to #1888)
TDD-first: 21 tests added covering exit_ticket clamping/fail-closed,
validate_student_sentence suggestion clearing, and grade_story_structure
fill_blank/checkbox/display scoring rules. All 21 pass before and after split.

Split 688L monolith into 4 focused modules:
- ai_generation/exit_ticket.py     — MCQ schema, prompt, generate + grade stubs
- ai_generation/sentence_practice.py — example sentences + validate (suggestion clear)
- ai_generation/story_structure.py  — STORY_STRUCTURE_SCHEMA, generate + grade (fuzzy match)
- ai_generation/teacher_comment.py  — teacher comment schema, prompt, generate

ai_generation/__init__.py re-exports all public names for zero-diff backward compat.
ai_service.py wildcard import and all route-level imports unchanged.
